requirements.txt
numpy
tensorflow==1.15.2
dm-sonnet<2
tensorflow-probability<0.9.0
flask
